+++ Moonlimit.DroneAPI/Moonlimit.DroneAPI.Domain/Service/Generic/IService.cs
@@ -8,10 +8,10 @@ namespace Moonlimit.DroneAPI.Domain.Service
     public interface IService<Tv, Te>
     {
         IEnumerable<Tv> GetAll();
-        int Add(Tv obj);
-        int Update(Tv obj);
-        int Remove(int id);
-        Tv GetOne(int id);
+        Int64 Add(Tv obj);
+        Int64 Update(Tv obj);
+        Int64 Remove(Int64 id);
+        Tv GetOne(Int64 id);
         IEnumerable<Tv> Get(Expression<Func<Te, bool>> predicate);
     }
 }
