+++ Moonlimit.DroneAPI/Moonlimit.DroneAPI.Entity/Entity/DroneCom/PlannedRoute.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Moonlimit.DroneAPI.Entity.DroneCom
@@ -5,7 +6,7 @@ namespace Moonlimit.DroneAPI.Entity.DroneCom
     [Table("planned_routes", Schema = "DroneCom")]
     public class PlannedRoute : BaseEntity
     {
-        public int DroneId { get; set; }
+        public Int64 DroneId { get; set; }
         [Column(TypeName = "jsonb")]
         public string WaypointsJson { get; set; }
     }
